requirements.txt
# requirements.txt - UPDATED

# Core Data Science
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Financial Data
yfinance==0.2.33
ta==0.11.0

# LLM & Embeddings - UPDATED VERSIONS
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Vector Databases
chromadb>=0.4.22

# Web Scraping & News
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10

# NLP & Text Processing
sentence-transformers==2.2.2
tiktoken>=0.5.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2

# Data validation
pydantic>=2.5.0,<3.0.0